背景函数基本绘制完成

+++ chartviewlib/src/main/java/com/example/will/chartviewlib/ChartInfo/BackgroundInfo/IScaleInfo.java
@@ -74,5 +74,5 @@ public interface IScaleInfo {
      * @param min
      * @param max
      */
-    void setYRange(float min, float max);
+    void setYRange(int min, int max);
 }
